// src/app/model/student.ts
export interface Student {
  id: string
  firstname: string
  lastname: string
  email: string
  room: number
  university: string
  phone: string
  am: string
  dateuniversity: string
  datedorm: string
  comments: string
}


// {
//   "id": "aa702ae8-6055-11ed-9b6a-0242ac120002",
//   "firstName": "George",
//   "lastName": "Karampelas",
//   "email": "george.akrampeleas.265@gmail.com",
//   "room": "205",
//   "university": "AUEB",
//   "phone": "6945222237",
//   "am": "32900",
//   "dateUniversity": "26/01/2018",
//   "dateDorm": "26/2/2019",
//   "comments": "FUCKIN SHITTS"
// }